code/requirements.txt
numpy
matplotlib
torch
torchvision
opencv-python
pillow
deep_translator
